Fixed Errors and Added AccessKey Feature
Some incompatibility issue was stopping `console.log` from working
correctly. This was fixed by casting the ...data object to a JSON
string, and casting it back to object using JSON.parse.

Errors in transactions are now captured in the VS "Widget" channel
in the same way as logs. This allows to quickly see what is the
error that stopped the transaction from going.

Adding the error catching made me saw that a functionality for adding
keys was missing, this PR adds a simple AddKey method, so people can
use non-monetary transactions to interact with smart contracts.